requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==4.2.3
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
